requirements.txt
Django==5.0.1
psycopg[binary]>=3.1.0
Pillow>=10.4.0
django-crispy-forms==2.1
crispy-tailwind==1.0.3
python-decouple==3.8
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
asgiref==3.7.2
